teamitaly-preselection-ctf/rev03/src/webapp/requirements.txt
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
gunicorn==20.1.0
eventlet==0.36.1